+++ SugarMarket/src/main/java/com/finance/sugarmarket/app/dto/CreditCardDto.java
@@ -2,14 +2,14 @@ package com.finance.sugarmarket.app.dto;
 
 public class CreditCardDto {
 
-	private Integer id;
+	private Long id;
 	private String bankName;
 	private String creditCardName;
 	private Integer statementDate;
 	private Integer dueDate;
 	private String last4Digit;
 
-	public CreditCardDto(Integer id, String bankName, String creditCardName, Integer statementDate, Integer dueDate,
+	public CreditCardDto(Long id, String bankName, String creditCardName, Integer statementDate, Integer dueDate,
 			String last4Digit) {
 		this.id = id;
 		this.bankName = bankName;
@@ -22,11 +22,11 @@ public class CreditCardDto {
 	public CreditCardDto() {
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
