+++ SugarMarket/src/main/java/com/finance/sugarmarket/app/repo/MutualFundRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.finance.sugarmarket.app.model.MutualFund;
 
-public interface MutualFundRepo extends JpaRepository<MutualFund, Integer> {
+public interface MutualFundRepo extends JpaRepository<MutualFund, Long> {
 	public MutualFund findMutualFundBySchemeCode(String schemeCode);
 	public MutualFund findMutualFundBySchemeName(String schemeName);
 }
