+++ SugarMarket/src/main/java/com/finance/sugarmarket/auth/repo/RoleRepo.java
@@ -3,6 +3,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.finance.sugarmarket.auth.model.MFRole;
 
-public interface RoleRepo extends JpaRepository<MFRole, Integer>{
+public interface RoleRepo extends JpaRepository<MFRole, Long>{
 
 }
